kafka/requirements.txt
aiohttp==3.9.5
asyncio==3.4.3
kafka-python==2.0.2
python-dotenv==1.0.1